refactor(back): update health state processing and instruction key retrieval
